Fixing linting
